Improve code distribution
