+++ src/components/combo-box/combo-box.tsx
@@ -32,6 +32,11 @@ import { ChPopoverCustomEvent } from "../../components";
 import { ChPopoverAlign } from "../popover/types";
 import { focusComposedPath } from "../common/helpers";
 import { filterSubModel } from "./helpers";
+import {
+  findComboBoxLargestValue,
+  getComboBoxItemImageCustomVars
+} from "./utils";
+import { findNextSelectedIndex, findSelectedIndex } from "./navigation";
 
 const SELECTED_PART = "selected";
 const DISABLED_PART = "disabled";
@@ -65,106 +70,6 @@ type KeyDownWithFiltersEvents =
   | typeof KEY_CODES.ARROW_DOWN
   | typeof KEY_CODES.ENTER
   | typeof KEY_CODES.TAB;
-
-const SELECTED_VALUE_DOES_NOT_EXISTS: ComboBoxSelectedIndex = {
-  type: "not-exists"
-} as const;
-
-const isValidIndex = (array: any, index: number) =>
-  0 <= index && index < array.length;
-
-const findSelectedIndex = (
-  valueToItemInfo: Map<string, ComboBoxItemModelExtended>,
-  selectedValue: string | undefined
-): ComboBoxSelectedIndex => {
-  if (!selectedValue) {
-    return SELECTED_VALUE_DOES_NOT_EXISTS;
-  }
-
-  return (
-    valueToItemInfo.get(selectedValue)?.index ?? SELECTED_VALUE_DOES_NOT_EXISTS
-  );
-};
-
-const findNextSelectedIndex = (
-  model: ComboBoxModel,
-  currentIndex: ComboBoxSelectedIndex,
-  increment: 1 | -1,
-  hasFilters: boolean,
-  displayedValues: Set<string>
-): ComboBoxSelectedIndex => {
-  if (currentIndex.type === "not-exists") {
-    return SELECTED_VALUE_DOES_NOT_EXISTS;
-  }
-  const firstLevelIndex = currentIndex.firstLevelIndex;
-
-  if (currentIndex.type === "nested") {
-    let secondLevelIndex = currentIndex.secondLevelIndex + increment; // Start from the first valid index
-    const firstLevelItemItems = (model[firstLevelIndex] as ComboBoxItemGroup)
-      .items;
-
-    // Search in the nested level skipping disabled and not rendered items
-    while (
-      isValidIndex(firstLevelItemItems, secondLevelIndex) &&
-      (firstLevelItemItems[secondLevelIndex].disabled ||
-        (hasFilters &&
-          !displayedValues.has(firstLevelItemItems[secondLevelIndex].value)))
-    ) {
-      secondLevelIndex += increment;
-    }
-
-    // If the index is not after the end of the array, the new selected value
-    // was found
-    if (isValidIndex(firstLevelItemItems, secondLevelIndex)) {
-      return {
-        type: "nested",
-        firstLevelIndex: firstLevelIndex,
-        secondLevelIndex: secondLevelIndex
-      };
-    }
-  }
-
-  // At this point, either all items in the nested level were disabled or the
-  // "currentIndex" was not nested. In any case, we must check the next item
-  // in the first level
-  let nextFirstLevelIndex = firstLevelIndex + increment;
-
-  // Search for the next first level item that is not disabled and is not filtered
-  while (
-    isValidIndex(model, nextFirstLevelIndex) &&
-    (model[nextFirstLevelIndex].disabled ||
-      (hasFilters && !displayedValues.has(model[nextFirstLevelIndex].value)))
-  ) {
-    nextFirstLevelIndex += increment;
-  }
-
-  // With this flag, we also say that we are at the end of the combo-box
-  // and there isn't any new "next value" to select
-  if (!isValidIndex(model, nextFirstLevelIndex)) {
-    return SELECTED_VALUE_DOES_NOT_EXISTS;
-  }
-
-  const nestedLevel = (model[nextFirstLevelIndex] as ComboBoxItemGroup).items;
-
-  if (nestedLevel != null) {
-    return findNextSelectedIndex(
-      model,
-      {
-        type: "nested",
-        firstLevelIndex: nextFirstLevelIndex,
-        secondLevelIndex: increment === 1 ? -1 : nestedLevel.length // The algorithm will sum 1 (or -1) to the start index
-      },
-      increment,
-      hasFilters,
-      displayedValues
-    );
-  }
-
-  return {
-    type: "first-level",
-    firstLevelIndex: nextFirstLevelIndex
-  };
-};
 
 type ImmediateFilter = "immediate" | "debounced" | undefined;
 
@@ -539,7 +444,6 @@ export class ChComboBoxRender
     if (!this.suggest) {
       this.filter = this.#currentValueCaption;
     }
-
     // Update form value
     this.internals.setFormValue(newValue);
   }
@@ -560,26 +464,7 @@ export class ChComboBoxRender
   @Event() input: EventEmitter<string>;
 
   #findLargestValue = (model: ComboBoxModel) => {
-    this.#largestValue = "";
-    let largestValueLength = 0;
-
-    model.forEach((itemGroup: ComboBoxItemGroup) => {
-      const subItems = itemGroup.items;
-
-      if (itemGroup.caption.length > largestValueLength) {
-        this.#largestValue = itemGroup.caption;
-        largestValueLength = itemGroup.caption.length;
-      }
-
-      if (subItems != null) {
-        subItems.forEach(leaf => {
-          if (leaf.caption.length > largestValueLength) {
-            this.#largestValue = leaf.caption;
-            largestValueLength = leaf.caption.length;
-          }
-        });
-      }
-    });
+    this.#largestValue = findComboBoxLargestValue(model);
   };
 
   #scheduleFilterProcessing = (newImmediateFilter?: ImmediateFilter) => {
@@ -790,9 +675,8 @@ export class ChComboBoxRender
     this.#resizeObserver.observe(this.#maskRef ?? this.#selectRef);
   };
 
-  #updateBorderSizeRAF = () => {
+  #updateBorderSizeRAF = () =>
     this.#borderSizeRAF.perform(this.#updateBorderSize);
-  };
 
   #updateBorderSize = () => {
     // - - - - - - - - - - - - - DOM read operations - - - - - - - - - - - - -
@@ -949,23 +833,6 @@ export class ChComboBoxRender
     forceUpdate(this);
   };
 
-  #getItemImageCustomVars = (
-    item: ComboBoxItemModel,
-    hasImages: boolean,
-    hasStartImg: boolean,
-    hasEndImg: boolean
-  ) =>
-    hasImages
-      ? {
-          "--ch-combo-box-item-start-img": hasStartImg
-            ? `url("${item.startImgSrc}")`
-            : null,
-          "--ch-combo-box-item-end-img": hasEndImg
-            ? `url("${item.endImgSrc}")`
-            : null
-        }
-      : undefined;
-
   #isModelAlreadyFiltered = () => this.suggestOptions.alreadyProcessed === true;
 
   #customItemRender =
@@ -987,7 +854,7 @@ export class ChComboBoxRender
       const hasEndImg = !!item.endImgSrc;
       const hasImages = hasStartImg || hasEndImg;
 
-      const customVars = this.#getItemImageCustomVars(
+      const customVars = getComboBoxItemImageCustomVars(
         item,
         hasImages,
         hasStartImg,
@@ -1026,9 +893,7 @@ export class ChComboBoxRender
               style={customVars}
               disabled={isDisabled}
               type="button"
-              onClick={
-                canAddListeners ? this.#toggleExpandInGroup(itemGroup) : null
-              }
+              onClick={canAddListeners && this.#toggleExpandInGroup(itemGroup)}
             >
               <span
                 class={{
@@ -1104,14 +969,10 @@ export class ChComboBoxRender
           disabled={isDisabled}
           type="button"
           onClick={
-            canAddListeners
-              ? this.#selectValueAndClosePopover(item.value)
-              : null
+            canAddListeners && this.#selectValueAndClosePopover(item.value)
           }
           onMouseEnter={
-            canAddListeners
-              ? this.#updateCurrentSelectedValue(item.value)
-              : null
+            canAddListeners && this.#updateCurrentSelectedValue(item.value)
           }
         >
           {item.caption}
@@ -1262,7 +1123,7 @@ export class ChComboBoxRender
 
     const hasStartImg = currentItemInInput && !!currentItemInInput.startImgSrc;
 
-    const customVars = this.#getItemImageCustomVars(
+    const customVars = getComboBoxItemImageCustomVars(
       currentItemInInput,
       hasStartImg,
       hasStartImg,
